Created IBO for the Terrain, and an option to not use IBO when rendering it (because of the different texture coordinates).

+++ include/mesh/AbstractMesh.h
@@ -41,7 +41,7 @@ protected:
 	glm::vec3* tangents;
 	glm::vec3* bitangents;
 
-	vector<GLubyte> indices;
+	vector<GLuint> indices;
 
 	unsigned int id;
